requirements.txt
azure-storage-blob
azure-keyvault-secrets
azure-identity
azure-keyvault-keys
loguru
openpyxl
fastapi==0.110.0
python-multipart
uvicorn
python-multipart
PyMuPDF
mammoth==1.7.0
markdownify==0.11.6
pandas==2.2.1
pydantic_settings==2.2.1
pymssql==2.2.11
pypdf==4.1.0
SQLAlchemy==2.0.23
